Restructure Observe section
