+++ app/src/main/java/com/harrysoft/burstcoinexplorer/di/MainModule.java
@@ -3,7 +3,7 @@ package com.harrysoft.burstcoinexplorer.di;
 import com.harrysoft.burstcoinexplorer.accounts.ui.AccountsFragment;
 import com.harrysoft.burstcoinexplorer.events.ui.EventsFragment;
 import com.harrysoft.burstcoinexplorer.explore.ui.ExploreFragment;
-import com.harrysoft.burstcoinexplorer.observe.ObserveFragment;
+import com.harrysoft.burstcoinexplorer.observe.ui.ObserveFragment;
 
 import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
