+++ app/src/main/java/com/harrysoft/burstcoinexplorer/main/MainActivity.java
@@ -27,7 +27,7 @@ import com.harrysoft.burstcoinexplorer.burst.entity.SearchRequestType;
 import com.harrysoft.burstcoinexplorer.burst.util.BurstUtils;
 import com.harrysoft.burstcoinexplorer.events.ui.EventsFragment;
 import com.harrysoft.burstcoinexplorer.explore.ui.ExploreFragment;
-import com.harrysoft.burstcoinexplorer.observe.ObserveFragment;
+import com.harrysoft.burstcoinexplorer.observe.ui.ObserveFragment;
 import com.harrysoft.burstcoinexplorer.router.ExplorerRouter;
 
 import java.math.BigInteger;
