+++ app/src/main/java/com/harrysoft/burstcoinexplorer/observe/util/OnRefreshRequestListener.java
@@ -1,8 +1,8 @@
-package com.harrysoft.burstcoinexplorer.observe.pieutils;
+package com.harrysoft.burstcoinexplorer.observe.util;
 
 import android.support.annotation.NonNull;
 
-import com.harrysoft.burstcoinexplorer.observe.ObserveSubFragment;
+import com.harrysoft.burstcoinexplorer.observe.ui.ObserveSubFragment;
 
 public interface OnRefreshRequestListener {
     void requestRefresh(@NonNull ObserveSubFragment sender);
