+++ app/src/main/java/com/harrysoft/burstcoinexplorer/observe/util/PieUtils.java
@@ -1,4 +1,4 @@
-package com.harrysoft.burstcoinexplorer.observe.pieutils;
+package com.harrysoft.burstcoinexplorer.observe.util;
 
 import android.content.Context;
 import android.graphics.Color;
@@ -13,7 +13,7 @@ import com.github.mikephil.charting.highlight.Highlight;
 import com.github.mikephil.charting.listener.OnChartValueSelectedListener;
 import com.github.mikephil.charting.utils.ColorTemplate;
 import com.harrysoft.burstcoinexplorer.R;
-import com.harrysoft.burstcoinexplorer.observe.ObserveNodeValueFormatter;
+import com.harrysoft.burstcoinexplorer.observe.ui.ObserveNodeValueFormatter;
 
 import java.util.ArrayList;
