+++ app/src/main/java/com/harrysoft/burstcoinexplorer/observe/util/RemovableLabelPieEntry.java
@@ -1,4 +1,4 @@
-package com.harrysoft.burstcoinexplorer.observe.pieutils;
+package com.harrysoft.burstcoinexplorer.observe.util;
 
 import android.text.TextUtils;
